requirements.txt
# Updated Requirements for Multimodal AI System

Flask==2.2.3
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.28.2
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
transformers==4.27.1
langchain==0.0.139
faiss-cpu==1.7.3
pymongo==4.3.3
redis==4.5.1
pytest==7.2.2
pytest-cov==4.0.0
black==23.1.0
isort==5.12.0
flake8==6.0.0
werkzeug==2.0.3
Pillow==9.4.0  # For image processing
